refactor: split military panel, map state shell, and CtIconAction (#3878)
Extract military unit panel build/dialog parts to match naval/civilian
pattern, move GameMapAreaStateLogic shell gating into a library part, and
split CtIconAction state chrome; update repo spec and bus allow-list.

Refs #3878
