+++ app/lib/features/game/flame/map_state/game_map_area_state_logic.dart
@@ -7,6 +7,7 @@ import 'game_map_area_civilian_draft_projection.dart';
 import 'game_map_area_fleet_draft_projection.dart';
 import 'game_map_area_province_action_states.dart';
 
+part 'game_map_area_state_logic_shell.dart';
 part 'game_map_area_state_logic_work_targets.dart';
 part 'game_map_area_state_logic_draft_projection.dart';
 part 'game_map_area_state_logic_province_actions.dart';
@@ -34,12 +35,8 @@ class ShellEntryAutoCenter {
 /// `GameMapArea` widget, tests, and the order-suggestions SPEC pointer
 /// (`SPEC/program/order-suggestions.md` § Authoritative pipeline).
 class GameMapAreaStateLogic {
-  /// Full turn resolution is a no-op once military [Game.victory] is set or the
-  /// campaign calendar cap has been reached ([Game.calendarCampaignHalted]).
-  /// SPEC/game/victory.md § UI blocking.
-  static bool allowsFullTurnResolution(ct_models.Game game) {
-    return !game.calendarCampaignHalted && game.victory == null;
-  }
+  static bool allowsFullTurnResolution(ct_models.Game game) =>
+      GameMapAreaStateLogicShell.allowsFullTurnResolution(game);
 
   static const ({bool showIcon, bool enabled, bool hasExplorerUnits})
   kHiddenExplorerInlineActionState =
@@ -48,44 +45,26 @@ class GameMapAreaStateLogic {
   kHiddenBuilderInlineActionState =
       GameMapAreaProvinceActionStates.kHiddenBuilderInlineActionState;
 
-  static int regionIndexFromWorldRegionId(String regionId) {
-    if (regionId == kRegionNewWorld) return 1;
-    return 0; // oldWorld (default)
-  }
+  static int regionIndexFromWorldRegionId(String regionId) =>
+      GameMapAreaStateLogicShell.regionIndexFromWorldRegionId(regionId);
 
-  /// Resolves the in-game shell auto-center target for [currentPlayerId].
-  ///
-  /// Returns `null` when auto-center must be skipped: [currentPlayerId] is
-  /// `null` (global observe has no viewing player) or that player has no
-  /// `capitalTile`. SPEC/ui/empire-overview.md § Initial map viewport.
   static ShellEntryAutoCenter? resolveShellEntryAutoCenter({
     required ct_models.Game game,
     required String? currentPlayerId,
-  }) {
-    if (currentPlayerId == null) {
-      return null;
-    }
-    final capital = game.playerById(currentPlayerId)?.capitalTile;
-    if (capital == null) {
-      return null;
-    }
-    return ShellEntryAutoCenter(
-      tileKey: capital.toTileKey(),
-      regionIndex: regionIndexFromWorldRegionId(capital.regionId),
-    );
-  }
+  }) =>
+      GameMapAreaStateLogicShell.resolveShellEntryAutoCenter(
+        game: game,
+        currentPlayerId: currentPlayerId,
+      );
 
-  /// Work-target tile translation hook for assignment flows.
-  ///
-  /// Civilian draft projection and locate use exact assigned tile keys for every
-  /// work target, so no target-specific tile normalization is applied here.
   static String translateWorkTargetTileKey({
     required String tileKey,
     required String workTarget,
-  }) {
-    if (workTarget.isEmpty) return tileKey;
-    return tileKey;
-  }
+  }) =>
+      GameMapAreaStateLogicShell.translateWorkTargetTileKey(
+        tileKey: tileKey,
+        workTarget: workTarget,
+      );
 
   static const Set<String> kCacheFirstWorkTargets =
       GameMapAreaStateLogicWorkTargets.kCacheFirstWorkTargets;
